Split maintenance into jobs/ (recurring) and migrations/ (one-shot)
Separate "runs forever" from "run once": jobs/maintenance.ts holds the recurring
prunes (events / shared events / rate-limits) + the account-deletion purge;
migrations/backfills.ts holds the one-shot migrations (clearEventAttendees,
migratePublicCalendarsToShared, purgeNonSharedSharedEvents). Root maintenance.ts
re-exports both, keeping every internal.maintenance.* path fixed — verified by a
clean convex dev deploy (crons reference those paths) and the purgeUserData
itest. Full suite green.
